Moved time related methods to new DateTime class
