+++ src/com/jboby93/markovbot/App.java
@@ -40,8 +40,8 @@ public class App {
 	}
 
 	public static void main(String args[]) throws IOException {
-		startTimeString = Tools.getDateString();
-		startTime = Tools.getUNIXTimestamp();
+		startTimeString = DateTime.getDateString();
+		startTime = DateTime.getUNIXTimestamp();
 		reader = new Scanner(System.in);
 
 		about();
@@ -220,7 +220,7 @@ public class App {
 					} else {
 						switch (cmd[1]) {
 						case "time":
-							Logger.debug("getUNIXTimestamp() returned " + Tools.getUNIXTimestamp());
+							Logger.debug("getUNIXTimestamp() returned " + DateTime.getUNIXTimestamp());
 							break;
 						default:
 							System.out.println("Invalid debug command");
